AI Agent - Azure Codes/Demo2/requirements.txt
azure-ai-projects>=2.0.0a20250915020
azure-ai-agents
azure-identity
python-dotenv